Refactor release diff code
Move Helm Chart Release diff code into cmd package.
